requirements.txt
Adyen==8.0.0
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.6.0
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
autobahn==23.1.2
Automat==22.10.0
backcall==0.2.0
beautifulsoup4==4.12.0
binaryornot==0.4.4
black==23.1.0
bleach==6.0.0
boto3==1.26.79
botocore==1.29.79
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
comm==0.1.3
constantly==15.1.0
cookiecutter==2.1.1
coverage==7.1.0
cryptography==39.0.1
daphne==4.0.0
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
Django==4.1.7
django-browser-reload==1.7.0
django-cors-headers==3.14.0
django-filter==22.1
django-mjml==1.0
django-ninja==0.21.0
django-redis==5.2.0
django-rest-knox==4.2.0
django-storages==1.13.2
django-tailwind==3.5.0
djangochannelsrestframework==1.1.0
djangorestframework==3.14.0
executing==1.2.0
fastjsonschema==2.16.3
filelock==3.10.0
fqdn==1.5.1
frozenlist==1.3.3
googlemaps==4.10.0
gunicorn==20.1.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.4
importlib-metadata==6.0.0
incremental==22.10.0
ipykernel==6.22.0
ipython==8.11.0
ipython-genutils==0.2.0
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
jinja2-time==0.2.0
jmespath==1.0.1
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter_client==8.1.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
keyboard==0.13.5
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mistune==2.0.5
msgpack==1.0.5
multidict==6.0.4
mypy-extensions==1.0.0
nbclassic==0.5.3
nbclient==0.7.2
nbconvert==7.2.10
nbformat==5.8.0
nest-asyncio==1.5.6
notebook==6.5.3
notebook_shim==0.2.2
numpy==1.24.2
openai==0.27.4
packaging==23.0
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.11.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.1.1
priority==1.3.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==4.21.12
psutil==5.9.4
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.10.4
pycparser==2.21
pydantic==1.10.7
Pygments==2.14.0
pyOpenSSL==23.0.0
pyperclip==1.8.2
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
python-slugify==8.0.1
pytz==2022.7.1
pywatchman==1.4.1
PyYAML==6.0
pyzmq==25.0.2
redis==4.5.4
requests==2.28.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
s3transfer==0.6.0
scipy==1.10.1
Send2Trash==1.8.0
service-identity==21.1.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4
sqlparse>=0.4.4
stack-data==0.6.2
terminado==0.17.1
text-unidecode==1.3
tinycss2==1.2.1
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.5.0
uri-template==1.2.0
urllib3==1.26.14
virtualenv==20.21.0
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
yarl==1.8.2
zipp==3.12.0
zope.interface==5.5.2
